Code Clean Up
Separated some of my code and placed it into a module as a should have from the start.
